Rename Azure-specific files
