+++ crates/datadog-metrics-collector/src/azure_instance.rs
