+++ crates/datadog-metrics-collector/src/azure_tags.rs
